requirements.txt
psycopg[binary]==3.1.18
python-dotenv==1.0.0
beautifulsoup4==4.12.2
requests==2.31.0
pandas==1.5.3
numpy==1.24.3
pyarrow==13.0.0
sqlalchemy==2.0.44